refactor: layout을 common으로 위치 변경
